error: Set traces to stderr
Since standard streams in BetterC have been fixed on Windows for a while, switch to stderr for tracing purposes.

This also adds module names into trace outputs. There was a note that __MODULE__ and __FILE__ are broken, but the lack of details lead me to think it's no longer happening. Haven't notice this sort of bug in a while anyway.